requirements.txt
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.0.0
Pillow>=9.0.0
jupyter>=1.0.0
